backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-environ>=0.11
drf-spectacular>=0.27
mysqlclient>=2.2
selenium>=4.15
beautifulsoup4>=4.12
chromadb>=0.4
sentence-transformers>=2.2
groq>=0.4
openai>=1.0
gunicorn>=21.2
psycopg2-binary>=2.9
djangorestframework-simplejwt>=5.3
